hotfix UITest and arch
