+++ ImageFeed/Presenters/Profile/ProfileViewPresenter.swift
@@ -5,11 +5,10 @@
 //  Created by Artem Krasnov on 26.02.2024.
 //
 
-import UIKit
+import Foundation
 
 protocol ProfileViewPresenterProtocol {
     var view: ProfileViewControllerProtocol? {get set}
-    func updateAvatar()
     func startObservingProfileImageChanges()
     func setupGradientAnimation()
     func didLogout()
@@ -19,19 +18,6 @@ final class ProfileViewPresenter: ProfileViewPresenterProtocol {
     weak var view: ProfileViewControllerProtocol?
     var profileImageServiceObserver: NSObjectProtocol?
     var router: LogoutRouterProtocol = LogoutRouter()
-
-    func updateAvatar() {
-         guard
-             let view = view,
-             let avatarURL = ProfileImageService.shared.avatarURL,
-             let url = URL(string: avatarURL)
-         else { return }
-
-         view.logoImageView.kf.setImage(
-             with: url,
-             placeholder: UIImage(named: "User")
-         )
-     }
 
     func setupGradientAnimation() {
         guard let view = view else { return }
@@ -46,7 +32,7 @@ final class ProfileViewPresenter: ProfileViewPresenterProtocol {
                 queue: .main
             ) { [weak self] _ in
                 guard let self = self else {return}
-                updateAvatar()
+                view?.updateAvatar()
             }
     }
